Fixed index.html
The button didn't work. Turned out, when I was trying to add a feature but didn't have time, I made a new js function and assigned the button click event to execute that function. And, that function....had nothing in it. So that's why it wasn't working.  But, it should work now!